+++ extendr-api/src/wrapper/expr.rs
@@ -1,19 +1,23 @@
 use super::*;
 
-#[derive(Debug, PartialEq, Clone)]
-pub struct Expression {
+#[derive(PartialEq, Clone)]
+pub struct Expressions {
     pub(crate) robj: Robj,
 }
 
-impl Expression {
-    /// Wrapper for creating Expression (EXPRSXP) objects.
+impl Expressions {
+    /// Wrapper for creating Expressions (EXPRSXP) objects.
+    pub fn new() -> Self {
+        Expressions::from_values([Robj::from(()); 0])
+    }
+
+    /// Wrapper for creating Expressions (EXPRSXP) objects.
     /// ```
     /// use extendr_api::prelude::*;
     /// test! {
-    ///     let expr = r!(Expression::from_values(&[r!(0), r!(1), r!(2)]));
-    ///     assert_eq!(expr.is_expression(), true);
+    ///     let expr = r!(Expressions::from_values(&[r!(0), r!(1), r!(2)]));
+    ///     assert_eq!(expr.is_expressions(), true);
     ///     assert_eq!(expr.len(), 3);
-    ///     assert_eq!(format!("{:?}", expr), r#"r!(Expression::from_values([r!(0), r!(1), r!(2)]))"#);
     /// }
     /// ```
     pub fn from_values<V>(values: V) -> Self
@@ -30,5 +34,19 @@ impl Expression {
     /// Return an iterator over the values of this expression list.
     pub fn values(&self) -> ListIter {
         ListIter::from_parts(self.robj.clone(), 0, self.robj.len())
+    }
+}
+
+impl std::default::Default for Expressions {
+    fn default() -> Self {
+        Expressions::new()
+    }
+}
+
+impl std::fmt::Debug for Expressions {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Expressions")
+            .field("values", &self.values())
+            .finish()
     }
 }